refactor(meeting): extract FailedMeetingStore + TranscriptionQueueCoordinator from MeetingSessionController
Pure code motion, no behavior change (codebase audit 2026-07-08 wave 2, spec
W2-B). MeetingSessionController.swift (3684 lines) owned two coherent
subsystems that were fully inlined:

- FailedMeetingStore.swift: FailedMeetingItem + all failed-meeting
  queue/persistence/retry bookkeeping (retryFailedMeeting, dismissFailedMeeting,
  deleteFailedMeeting, preserveFailedMeetingForRetry,
  refreshTimedOutFailedMeetingAudio, refreshFailedMeetings computation,
  scheduleFailedAudioCompression).
- TranscriptionQueueCoordinator.swift: QueuedTranscriptionJob (+ Kind),
  BackgroundTranscriptionWorkSnapshot, and the queue enqueue/dispatch/recovery
  logic around taskManager (enqueueTranscriptionJob, startQueuedTranscription,
  ensureModelsReadyForQueuedTranscription, runPreparedQueuedTranscription,
  failQueuedTranscriptionJobAfterModelRecovery, handleBackgroundTranscription-
  WorkChanged, etc.).

Both are plain owned objects (`let failedMeetingStore`, `let
transcriptionQueue`), holding an `unowned` back-reference to the controller
and reaching into it for shared state (taskManager, sttAdapter, diarization,
downloader, sttRouter, live-sidecar bookkeeping) rather than duplicating it.
The controller keeps the 6-state state machine and the full @Published
surface — `setState`/`setDisplayStatus` are the two narrow callbacks the
owned objects use to drive `state`/`displayStatus` transitions, matching the
"callbacks back into the controller" design in the spec. `FailedMeetingItem`,
`QueuedTranscriptionJob`, and `BackgroundTranscriptionWorkSnapshot` stay
resolvable as `MeetingSessionController.X` via typealiases so no UI or Core
call site changes. retryFailedMeeting/dismissFailedMeeting/deleteFailedMeeting
keep their exact public signatures as thin forwarders.

Several `private` members (sttRouter, sttAdapter, diarization, downloader,
failedManager, hasBackgroundTranscriptionWork, isCaptureSessionActive,
baseDiagnosticsContext, boolString, finishLiveCodexSession,
refreshFailedMeetings, and a handful of controller-owned var fields) widen to
internal so the two new same-module files can reach them — no public API
changes outside the Meeting subsystem.

Deviations from the spec:
- Controller lands at 3113 lines, not <2200. The two subsystems as scoped
  by the spec total ~570 lines of real motion; hitting <2200 would require
  pulling in code (state machine, live-sidecar handoff, telemetry) the spec
  explicitly said to leave in the controller this wave, given the deep
  coupling between all of it (shared @Published state, shared diagnostics
  helpers). Flagging for the merge coordinator rather than expanding scope.
- StartTrigger/StopReason/RecordingCancelReason/TranscriptionCancelReason
  stayed nested in the controller (not moved to a MeetingSessionTypes.swift)
  per the spec's "if referenced outside" condition — grep found no call
  sites outside MeetingSessionController.swift for any of the four.
- `failedMeetingStore`/`transcriptionQueue` are implicitly-unwrapped
  optionals (`FailedMeetingStore!`), not plain `let`: Swift's definite-
  initialization checker rejects passing `self` into either type's `init`
  (both hold `unowned let controller`) before the property being assigned
  is itself set — even for the last two properties in the initializer.
  They're constructed unconditionally, immediately, before any other code
  runs, and are `private(set)` so nothing outside init ever reassigns them.
- Tests/RepoCommandContractTests.swift and
  Tests/AuditRegressionCoverageContractTests.swift are golden-master-style
  contracts that assert exact source text/order inside
  MeetingSessionController.swift; ~10 suites read code that intentionally
  moved this wave. Retargeted `readRepoTextFile` calls and `sourceSlice`
  markers to the new files where the underlying invariant moved with the
  code; left assertions untouched wherever plain substring matching still
  held after the `controller.`/`transcriptionQueue.`/`failedMeetingStore.`
  prefixes were added. No behavioral test files needed changes.
- run-tests.sh's APP_SOURCES and Tests/FastTests.manifest were not touched:
  MeetingSessionController.swift itself was never listed in either (it's
  Core-wiring code outside the fast-test compilation closure), so the two
  new files that split out of it don't need tracking there either.

Verified: `bash build.sh` (App target, full signed build) succeeds.
`bash scripts/entrypoints/run-tests.sh` — 12863 tests, 12863 passed, 0
failed, including the full RepoCommandContractTests, Audit-
RegressionCoverageContractTests, FailedMeetingPresentationTests, and
AnalyticsEventPolicyTests suites.
